refactor:structure
