+++ app/sections/author_page/AuthorBlogs.module.scss
@@ -1,5 +1,5 @@
-@import "./../../styles/mixin";
-@import "./../../styles/container";
+@import "./../../../styles/mixin";
+@import "./../../../styles/container";
 
 .articles {
   display: flex;
